# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartSolutions CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List all proposals
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of proposals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create a new proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalRequest"
      responses:
        "201":
          description: Proposal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get a proposal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    put:
      operationId: updateProposal
      tags: [proposals]
      summary: Update a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalRequest"
      responses:
        "200":
          description: Proposal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete a proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Proposal deleted
  /proposals/{id}/send:
    post:
      operationId: sendProposal
      tags: [proposals]
      summary: Send a proposal by email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendProposalRequest"
      responses:
        "200":
          description: Proposal sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        benefits:
          type: string
        objections:
          type: string
        salesArguments:
          type: string
        price:
          type: string
          nullable: true
        pricingType:
          type: string
        attachmentUrl:
          type: string
          nullable: true
        proposalTemplate:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - shortDescription
        - fullDescription
        - benefits
        - objections
        - salesArguments
        - pricingType
        - active
        - createdAt
        - updatedAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        benefits:
          type: string
        objections:
          type: string
        salesArguments:
          type: string
        price:
          type: string
          nullable: true
        pricingType:
          type: string
        attachmentUrl:
          type: string
          nullable: true
        proposalTemplate:
          type: string
          nullable: true
        active:
          type: boolean
      required:
        - name
        - category
        - shortDescription
        - fullDescription
        - benefits
        - objections
        - salesArguments
        - pricingType
        - active
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        benefits:
          type: string
        objections:
          type: string
        salesArguments:
          type: string
        price:
          type: string
          nullable: true
        pricingType:
          type: string
        attachmentUrl:
          type: string
          nullable: true
        proposalTemplate:
          type: string
          nullable: true
        active:
          type: boolean
    Proposal:
      type: object
      properties:
        id:
          type: integer
        proposalNumber:
          type: string
        clientId:
          type: integer
        productId:
          type: integer
          nullable: true
        clientEmail:
          type: string
        amount:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
        sentAt:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        attachmentUrl:
          type: string
          nullable: true
        ownerName:
          type: string
        updatedAt:
          type: string
        clientName:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        productName:
          type: string
          nullable: true
      required:
        - id
        - proposalNumber
        - clientId
        - clientEmail
        - status
        - createdAt
        - ownerName
        - updatedAt
    CreateProposalRequest:
      type: object
      properties:
        clientId:
          type: integer
        productId:
          type: integer
          nullable: true
        clientEmail:
          type: string
        amount:
          type: string
          nullable: true
        status:
          type: string
        validUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        ownerName:
          type: string
      required:
        - clientId
        - clientEmail
        - status
        - ownerName
    UpdateProposalRequest:
      type: object
      properties:
        clientId:
          type: integer
        productId:
          type: integer
          nullable: true
        clientEmail:
          type: string
        amount:
          type: string
          nullable: true
        status:
          type: string
        validUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        ownerName:
          type: string
        attachmentUrl:
          type: string
          nullable: true
    SendProposalRequest:
      type: object
      properties:
        subject:
          type: string
        message:
          type: string
      required:
        - subject
        - message
...