# lib/api-zod/src/generated/types/createProductRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartSolutions CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductRequest {
  name: string;
  category: string;
  shortDescription: string;
  fullDescription: string;
  benefits: string;
  objections: string;
  salesArguments: string;
  price?: string | null;
  pricingType: string;
  attachmentUrl?: string | null;
  proposalTemplate?: string | null;
  active: boolean;
}