# lib/api-zod/src/generated/types/createProposalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartSolutions CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProposalRequest {
  clientId: number;
  productId?: number | null;
  clientEmail: string;
  amount?: string | null;
  status: string;
  validUntil?: string | null;
  notes?: string | null;
  ownerName: string;
}